frameworks/opt/hardware: Add DisplayColorCalibration class, leave the other two empty for now
Display is different on a ton of devices, with very little in common. I've basically
hacked together a DisplayColorCalibration class that should work for most scenarios. Really
not sure if I want to do the same with the others... Or just have device maintainers add
them in on their own.